Resolves warnings in EditDeck
Most warnings resolved, but there are a couple issues with the public key in our deck prop.
(1) public is a reserved word in strict mode
(2) public is showing as an unused state property because its value is not being used in the render function. We should probably refactor so that public is being used in the render and not changed through the handleCardChange exclusively